requirements.txt
Django==2.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
importlib-metadata==3.10.0
install==1.3.4
Markdown==3.3.4
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
